+++ src/man/motion/bhwalk/Tools/ForwardKinematic.h
@@ -20,15 +20,15 @@ public:
     MassCalibrationBH::Limb shoulder = left ? MassCalibrationBH::shoulderLeft : MassCalibrationBH::shoulderRight;
     JointDataBH::Joint arm0 = left ? JointDataBH::LShoulderPitch : JointDataBH::RShoulderPitch;
 
-    limbs[shoulder + 0] = Pose3DBH(robotDimensions.armOffset.x, robotDimensions.armOffset.y * -sign, robotDimensions.armOffset.z)
+    limbs[shoulder + 0] = Pose3DBH(robotDimensions.armOffset.x, robotDimensions.armOffset.y * (float)-sign, robotDimensions.armOffset.z)
                           .rotateY(-joints.angles[arm0 + 0]);
     limbs[shoulder + 1] = Pose3DBH(limbs[shoulder + 0])
-                          .rotateZ(joints.angles[arm0 + 1] * -sign);
+                          .rotateZ(joints.angles[arm0 + 1] * (float)-sign);
     limbs[shoulder + 2] = Pose3DBH(limbs[shoulder + 1])
-                          .translate(robotDimensions.upperArmLength, robotDimensions.yElbowShoulder * -sign, 0)
-                          .rotateX(joints.angles[arm0 + 2] * -sign);
+                          .translate(robotDimensions.upperArmLength, robotDimensions.yElbowShoulder * (float)-sign, 0)
+                          .rotateX(joints.angles[arm0 + 2] * (float)-sign);
     limbs[shoulder + 3] = Pose3DBH(limbs[shoulder + 2])
-                          .rotateZ(joints.angles[arm0 + 3] * -sign);
+                          .rotateZ(joints.angles[arm0 + 3] * (float)-sign);
   }
 
   static void calculateLegChain(bool left, const JointDataBH& joints, const RobotDimensionsBH& robotDimensions, const MassCalibrationBH& massCalibration, Pose3DBH limbs[MassCalibrationBH::numOfLimbs])
@@ -37,12 +37,12 @@ public:
     MassCalibrationBH::Limb pelvis = left ? MassCalibrationBH::pelvisLeft : MassCalibrationBH::pelvisRight;
     JointDataBH::Joint leg0 = left ? JointDataBH::LHipYawPitch : JointDataBH::RHipYawPitch;
 
-    limbs[pelvis + 0] =  Pose3DBH(0, robotDimensions.lengthBetweenLegs / 2.0f * -sign, 0)
-                         .rotateX(-pi_4 * sign)
-                         .rotateZ(joints.angles[leg0 + 0] * sign)
-                         .rotateX(pi_4 * sign);
+    limbs[pelvis + 0] =  Pose3DBH(0, robotDimensions.lengthBetweenLegs / 2.0f * (float)-sign, 0)
+                         .rotateX(-pi_4 * (float)sign)
+                         .rotateZ(joints.angles[leg0 + 0] * (float)sign)
+                         .rotateX(pi_4 * (float)sign);
     limbs[pelvis + 1] = Pose3DBH(limbs[pelvis + 0])
-                        .rotateX(joints.angles[leg0 + 1] * sign);
+                        .rotateX(joints.angles[leg0 + 1] * (float)sign);
     limbs[pelvis + 2] = Pose3DBH(limbs[pelvis + 1])
                         .rotateY(joints.angles[leg0 + 2]);
     limbs[pelvis + 3] = Pose3DBH(limbs[pelvis + 2])
@@ -52,7 +52,7 @@ public:
                         .translate(0, 0, -robotDimensions.lowerLegLength)
                         .rotateY(joints.angles[leg0 + 4]);
     limbs[pelvis + 5] = Pose3DBH(limbs[pelvis + 4])
-                        .rotateX(joints.angles[leg0 + 5] * sign);
+                        .rotateX(joints.angles[leg0 + 5] * (float)sign);
   }
 
   static void calculateHeadChain(const JointDataBH& joints, const RobotDimensionsBH& robotDimensions, const MassCalibrationBH& massCalibration, Pose3DBH limbs[MassCalibrationBH::numOfLimbs])
