+++ src/man/motion/bhwalk/Tools/Math/Common.h
@@ -115,7 +115,7 @@ template <class V> inline V fromDegrees(const V& degrees) {return degrees * V(pi
 template <class V> inline V normalizeBH(const V& data)
 {
   if(data < V(pi) && data >= -V(pi)) return data;
-  V ndata = data - ((int)(data / V(pi2))) * V(pi2);
+  V ndata = data - (float)((int)(data / V(pi2))) * V(pi2);
   if(ndata >= V(pi))
   {
     ndata -= V(pi2);
